Code refactored, moved functions into functions.js, added a default route
