+++ .gitignore
@@ -1,2 +1,5 @@
 /node_modules/
 /log/
+q
+history/
+dolar.json
